test: split describe-feature.spec.ts in other files
